update proj1b 2

+++ proj1b/OffByN.java
@@ -6,6 +6,6 @@ public class OffByN implements CharacterComparator {
     }
 
     public boolean equalChars(char x, char y) {
-        return Math.abs((int)x - y) == n;
+        return Math.abs(x - y) == n;
     }
 }
