+++ proj1b/OffByOne.java
@@ -1,7 +1,7 @@
 public class OffByOne implements CharacterComparator {
     @Override
     public boolean equalChars(char x, char y) {
-        return Math.abs((int)(x - y)) == 1;
+        return Math.abs(x - y) == 1;
     }
 
 }
